// arcadia.game.core.worker/src/modules/coreClient/enum/core.message.type.ts
export enum CoreMessage {
  IDLE_TIMEOUT = 'idleTimeout',
  GRACE_TIMEOUT = 'graceTimeout',
  QUEUE_CHANGE_OFFERS = 'queueChangeOffers',
  ENGAGE_TIMEOUT = 'engageTimeout',
  JACKPOT_RELOGIN_TIMEOUT = 'jackpotReloginTimeout',
  PHANTOM_DELAY = 'phantomDelay',
  PING_OUTDATED = 'pingOutdated',
  ROUND_END_DELAY = 'roundEndDelay',
  TERMINATE_VIEWERS = 'terminateViewers',
}